app/requirements.txt
fastapi
starlette
mysql-connector-python
hypercorn
requests
